+dialogs/HistoryUI - deployed history component from SerialUI.
* plan to use it with plugins
